api/requirements.txt
PyMuPDF==1.26.5
fastapi==0.120.4
uvicorn==0.38.0
python-multipart==0.0.20
python-dotenv==1.2.1
langchain-core==1.0.2
langchain_openai==1.0.1
langchain-community==0.4.1